PLH-python/requirements.txt
langchain>=0.1.0
langchain-together>=0.0.3
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=10.0.0 